+++ previous-versions/files/codatfiles.go
@@ -149,9 +149,9 @@ func New(opts ...SDKOption) *CodatFiles {
 		sdkConfiguration: sdkConfiguration{
 			Language:          "go",
 			OpenAPIDocVersion: "3.0.0",
-			SDKVersion:        "0.26.0",
-			GenVersion:        "2.188.1",
-			UserAgent:         "speakeasy-sdk/go 0.26.0 2.188.1 3.0.0 github.com/codatio/client-sdk-go/previous-versions/files",
+			SDKVersion:        "0.26.1",
+			GenVersion:        "2.194.1",
+			UserAgent:         "speakeasy-sdk/go 0.26.1 2.194.1 3.0.0 github.com/codatio/client-sdk-go/previous-versions/files",
 		},
 	}
 	for _, opt := range opts {