+++ previous-versions/files/pkg/models/sdkerrors/schema.go
@@ -6,6 +6,7 @@ import (
 	"encoding/json"
 )
 
+// Schema - Your API request was not properly authorized.
 type Schema struct {
 	// `True` if the error occurred transiently and can be retried.
 	CanBeRetried *string `json:"canBeRetried,omitempty"`